Split JavaExecution class
